Fix openstack docs to state only name of the flavor is valid
The current documentation states you can provide both the name and id of
an openstack flavor size. This is not the case, when providing the id of
a flavor, teflo throws an exception stating it cannot find the flavor
size object.

This commit adjusts the methods docstring and the rst docs to reflect
only being able to provide the name.